requirements.txt
requests>=2.25.0
pyaudio>=0.2.11
pygame>=2.0.0
psutil>=5.8.0
openai>=1.40.0          # LLM/视觉模型
pyautogui>=0.9.54       # 截屏分析
pvporcupine>=3.0.0      # 语音唤醒
python-dotenv>=1.0.0    # 读取 .env

